LPS-192496 Merge ClusterConfigurationModelIndexer with ConfigurationModelIndexer and Register IdentifiableOSGiService manually
