--- modules/apps/configuration-admin/configuration-admin-web/src/main/java/com/liferay/configuration/admin/web/internal/search/ClusterConfigurationModelIndexer.java
@@ -1,120 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.configuration.admin.web.internal.search;
-
-import com.liferay.configuration.admin.web.internal.model.ConfigurationModel;
-import com.liferay.portal.kernel.cluster.ClusterExecutor;
-import com.liferay.portal.kernel.cluster.ClusterMasterExecutor;
-import com.liferay.portal.kernel.cluster.ClusterMasterTokenTransitionListener;
-import com.liferay.portal.kernel.cluster.ClusterRequest;
-import com.liferay.portal.kernel.module.framework.service.IdentifiableOSGiService;
-import com.liferay.portal.kernel.module.framework.service.IdentifiableOSGiServiceUtil;
-import com.liferay.portal.kernel.util.MethodHandler;
-import com.liferay.portal.kernel.util.MethodKey;
-
-import java.util.Collection;
-
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.util.tracker.BundleTracker;
-
-/**
- * @author Tina Tian
- */
-@Component(
-	service = {
-		ClusterConfigurationModelIndexer.class, IdentifiableOSGiService.class
-	}
-)
-public class ClusterConfigurationModelIndexer
-	implements IdentifiableOSGiService {
-
-	@Override
-	public String getOSGiServiceIdentifier() {
-		return ClusterConfigurationModelIndexer.class.getName();
-	}
-
-	@Activate
-	protected void activate() {
-		if (_clusterExecutor.isEnabled()) {
-			_configurationModelsClusterMasterTokenTransitionListener =
-				new ConfigurationModelsClusterMasterTokenTransitionListener();
-
-			_clusterMasterExecutor.addClusterMasterTokenTransitionListener(
-				_configurationModelsClusterMasterTokenTransitionListener);
-		}
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		if (_configurationModelsClusterMasterTokenTransitionListener != null) {
-			_clusterMasterExecutor.removeClusterMasterTokenTransitionListener(
-				_configurationModelsClusterMasterTokenTransitionListener);
-		}
-
-		_stopBundleTracker();
-	}
-
-	private static void _reset(String osgiServiceIdentifier) {
-		ClusterConfigurationModelIndexer clusterConfigurationModelIndexer =
-			(ClusterConfigurationModelIndexer)
-				IdentifiableOSGiServiceUtil.getIdentifiableOSGiService(
-					osgiServiceIdentifier);
-
-		clusterConfigurationModelIndexer._initialized = false;
-	}
-
-	private synchronized void _stopBundleTracker() {
-		if (_bundleTracker != null) {
-			_bundleTracker.close();
-
-			_bundleTracker = null;
-		}
-	}
-
-	private static final MethodKey _resetMethodKey = new MethodKey(
-		ClusterConfigurationModelIndexer.class, "_reset", String.class);
-
-	private BundleTracker<Collection<ConfigurationModel>> _bundleTracker;
-
-	@Reference
-	private ClusterExecutor _clusterExecutor;
-
-	@Reference
-	private ClusterMasterExecutor _clusterMasterExecutor;
-
-	private ConfigurationModelsClusterMasterTokenTransitionListener
-		_configurationModelsClusterMasterTokenTransitionListener;
-	private volatile boolean _initialized;
-
-	private class ConfigurationModelsClusterMasterTokenTransitionListener
-		implements ClusterMasterTokenTransitionListener {
-
-		@Override
-		public void masterTokenAcquired() {
-			_initialized = false;
-
-			ClusterRequest clusterRequest =
-				ClusterRequest.createMulticastRequest(
-					new MethodHandler(
-						_resetMethodKey, getOSGiServiceIdentifier()),
-					true);
-
-			clusterRequest.setFireAndForget(true);
-
-			_clusterExecutor.execute(clusterRequest);
-		}
-
-		@Override
-		public void masterTokenReleased() {
-			_stopBundleTracker();
-		}
-
-	}
-
-}
