+++ modules/apps/configuration-admin/configuration-admin-web/src/main/java/com/liferay/configuration/admin/web/internal/search/ConfigurationModelIndexer.java
@@ -14,12 +14,16 @@ import com.liferay.configuration.admin.web.internal.util.ResourceBundleLoaderPro
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClassDefinition;
+import com.liferay.portal.kernel.cluster.ClusterExecutor;
 import com.liferay.portal.kernel.cluster.ClusterMasterExecutor;
+import com.liferay.portal.kernel.cluster.ClusterMasterTokenTransitionListener;
+import com.liferay.portal.kernel.cluster.ClusterRequest;
 import com.liferay.portal.kernel.concurrent.NoticeableFuture;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.CompanyConstants;
+import com.liferay.portal.kernel.module.framework.service.IdentifiableOSGiService;
 import com.liferay.portal.kernel.module.framework.service.IdentifiableOSGiServiceUtil;
 import com.liferay.portal.kernel.resource.bundle.ResourceBundleLoader;
 import com.liferay.portal.kernel.search.BaseIndexer;
@@ -59,8 +63,10 @@ import javax.portlet.PortletResponse;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.BundleEvent;
+import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 import org.osgi.service.metatype.AttributeDefinition;
 import org.osgi.service.metatype.ObjectClassDefinition;
@@ -72,9 +78,10 @@ import org.osgi.util.tracker.BundleTrackerCustomizer;
  */
 @Component(
 	property = {"index.on.startup=false", "system.index=true"},
-	service = {ConfigurationModelIndexer.class, Indexer.class}
+	service = Indexer.class
 )
-public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
+public class ConfigurationModelIndexer
+	extends BaseIndexer<ConfigurationModel> implements IdentifiableOSGiService {
 
 	@Override
 	public String getClassName() {
@@ -104,6 +111,11 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 		catch (Exception exception) {
 			throw new SearchException(exception);
 		}
+	}
+
+	@Override
+	public String getOSGiServiceIdentifier() {
+		return ConfigurationModelIndexer.class.getName();
 	}
 
 	public BundleTracker<Collection<ConfigurationModel>> initialize() {
@@ -209,6 +221,17 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 		setStagingAware(false);
 
 		_bundleContext = bundleContext;
+
+		_serviceRegistration = _bundleContext.registerService(
+			IdentifiableOSGiService.class, this, null);
+
+		if (_clusterExecutor.isEnabled()) {
+			_configurationModelsClusterMasterTokenTransitionListener =
+				new ConfigurationModelsClusterMasterTokenTransitionListener();
+
+			_clusterMasterExecutor.addClusterMasterTokenTransitionListener(
+				_configurationModelsClusterMasterTokenTransitionListener);
+		}
 	}
 
 	@Override
@@ -246,6 +269,17 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 		fullBooleanQuery.add(searchQuery, BooleanClauseOccur.MUST);
 
 		return fullBooleanQuery;
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		if (_configurationModelsClusterMasterTokenTransitionListener != null) {
+			_clusterMasterExecutor.removeClusterMasterTokenTransitionListener(
+				_configurationModelsClusterMasterTokenTransitionListener);
+		}
+
+		_serviceRegistration.unregister();
+		_stopBundleTracker();
 	}
 
 	@Override
@@ -390,6 +424,15 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 		configurationModelIndexer._initialize();
 	}
 
+	private static void _reset(String osgiServiceIdentifier) {
+		ConfigurationModelIndexer configurationModelIndexer =
+			(ConfigurationModelIndexer)
+				IdentifiableOSGiServiceUtil.getIdentifiableOSGiService(
+					osgiServiceIdentifier);
+
+		configurationModelIndexer._initialized = false;
+	}
+
 	private void _addLocalizedText(
 		Document document, ResourceBundleLoader resourceBundleLoader,
 		List<TranslationHelper> translationHelpers) {
@@ -489,9 +532,7 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 				NoticeableFuture<Void> noticeableFuture =
 					_clusterMasterExecutor.executeOnMaster(
 						new MethodHandler(
-							_initializeMethodKey,
-							_clusterConfigurationModelIndexer.
-								getOSGiServiceIdentifier()));
+							_initializeMethodKey, getOSGiServiceIdentifier()));
 
 				try {
 					noticeableFuture.get();
@@ -515,16 +556,27 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 			configurationModel.getFactoryPid());
 	}
 
+	private synchronized void _stopBundleTracker() {
+		if (_bundleTracker != null) {
+			_bundleTracker.close();
+
+			_bundleTracker = null;
+		}
+	}
+
 	private static final Log _log = LogFactoryUtil.getLog(
 		ConfigurationModelIndexer.class);
 
 	private static final MethodKey _initializeMethodKey = new MethodKey(
 		ConfigurationModelIndexer.class, "_initialize", String.class);
+	private static final MethodKey _resetMethodKey = new MethodKey(
+		ConfigurationModelIndexer.class, "_reset", String.class);
 
 	private BundleContext _bundleContext;
+	private BundleTracker<Collection<ConfigurationModel>> _bundleTracker;
 
 	@Reference
-	private ClusterConfigurationModelIndexer _clusterConfigurationModelIndexer;
+	private ClusterExecutor _clusterExecutor;
 
 	@Reference
 	private ClusterMasterExecutor _clusterMasterExecutor;
@@ -534,6 +586,9 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 
 	@Reference(target = "(!(filter.visibility=*))")
 	private ConfigurationModelRetriever _configurationModelRetriever;
+
+	private ConfigurationModelsClusterMasterTokenTransitionListener
+		_configurationModelsClusterMasterTokenTransitionListener;
 
 	@Reference
 	private IndexStatusManager _indexStatusManager;
@@ -548,6 +603,8 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 
 	@Reference
 	private ResourceBundleLoaderProvider _resourceBundleLoaderProvider;
+
+	private ServiceRegistration<IdentifiableOSGiService> _serviceRegistration;
 
 	private static class TranslationHelper {
 
@@ -637,6 +694,31 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 
 		private final Map<String, Collection<ConfigurationModel>>
 			_configurationModelsMap;
+
+	}
+
+	private class ConfigurationModelsClusterMasterTokenTransitionListener
+		implements ClusterMasterTokenTransitionListener {
+
+		@Override
+		public void masterTokenAcquired() {
+			_initialized = false;
+
+			ClusterRequest clusterRequest =
+				ClusterRequest.createMulticastRequest(
+					new MethodHandler(
+						_resetMethodKey, getOSGiServiceIdentifier()),
+					true);
+
+			clusterRequest.setFireAndForget(true);
+
+			_clusterExecutor.execute(clusterRequest);
+		}
+
+		@Override
+		public void masterTokenReleased() {
+			_stopBundleTracker();
+		}
 
 	}
